requirements.txt
# Core PyTorch and ML libraries
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
numpy>=1.21.0
scipy>=1.7.0

# Data manipulation and analysis
pandas>=1.3.0
scikit-learn>=1.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Jupyter and notebooks
jupyter>=1.0.0
jupyterlab>=3.0.0
ipywidgets>=7.6.0
notebook>=6.4.0

# Development and testing
pytest>=6.2.0
pytest-cov>=2.12.0
black>=21.0.0
flake8>=3.9.0
isort>=5.9.0

# Documentation
sphinx>=4.0.0
sphinx-rtd-theme>=1.0.0
nbsphinx>=0.8.0

# Utilities
tqdm>=4.62.0
pillow>=8.3.0
requests>=2.26.0
pyyaml>=5.4.0

# Optional: Advanced libraries (comment out if not needed)
transformers>=4.20.0
datasets>=2.0.0
torchsummary>=1.5.0
torchmetrics>=0.9.0
lightning>=2.0.0

# Optional: Optimization libraries
optuna>=2.10.0
wandb>=0.12.0

# Optional: Computer Vision
opencv-python>=4.5.0
albumentations>=1.2.0

# Optional: NLP
nltk>=3.6.0
spacy>=3.4.0
tokenizers>=0.12.0

# Optional: Audio processing
librosa>=0.9.0
soundfile>=0.10.0